+++ packages/boolean/stories/BooleanEditor.stories.tsx
@@ -1,18 +1,19 @@
 import * as React from 'react';
 
+import { FormControl } from '@contentful/f36-components';
 import { ActionsPlayground, createFakeFieldAPI } from '@contentful/field-editor-test-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import * as Editor from '../src/BooleanEditor';
+import { BooleanEditor } from '../src/BooleanEditor';
 
-const meta: Meta<typeof Editor.BooleanEditor> = {
+const meta: Meta<typeof BooleanEditor> = {
   title: 'editors/Boolean',
-  component: Editor.BooleanEditor,
+  component: BooleanEditor,
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Editor.BooleanEditor>;
+type Story = StoryObj<typeof BooleanEditor>;
 
 export const Default: Story = {
   parameters: {
@@ -21,10 +22,10 @@ export const Default: Story = {
   render: () => {
     const [field, mitt] = createFakeFieldAPI();
     return (
-      <div>
-        <Editor.BooleanEditor field={field} isInitiallyDisabled={false} />
+      <FormControl as="fieldset" aria-label="Choose one">
+        <BooleanEditor field={field} isInitiallyDisabled={false} />
         <ActionsPlayground mitt={mitt} />
-      </div>
+      </FormControl>
     );
   },
 };