+++ packages/reference/src/entries/WrappedEntryCard/FetchingWrappedEntryCard.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { EntryCard } from '@contentful/f36-components';
-import { useLocalePublishStatus, useActiveLocales } from '@contentful/field-editor-shared';
+import { useLocalePublishStatus } from '@contentful/field-editor-shared';
 import { EntryProps } from 'contentful-management';
 import get from 'lodash/get';
 
@@ -25,6 +25,9 @@ export type EntryCardReferenceEditorProps = ReferenceEditorProps & {
   onMoveBottom?: () => void;
   renderCustomMissingEntityCard?: RenderCustomMissingEntityCard;
   isBeingDragged?: boolean;
+  activeLocales?: {
+    code: string;
+  }[];
 };
 
 async function openEntry(
@@ -65,7 +68,6 @@ export function FetchingWrappedEntryCard(props: EntryCardReferenceEditorProps) {
     [getEntityScheduledActions, props.entryId]
   );
   const localesStatusMap = useLocalePublishStatus(entry, props.sdk.locales);
-  const activeLocales = useActiveLocales(props.sdk);
 
   const size = props.viewType === 'link' ? 'small' : 'default';
   const { getEntity } = useEntityLoader();
@@ -147,7 +149,7 @@ export function FetchingWrappedEntryCard(props: EntryCardReferenceEditorProps) {
       isBeingDragged: props.isBeingDragged,
       useLocalizedEntityStatus: props.sdk.parameters.instance.useLocalizedEntityStatus,
       localesStatusMap,
-      activeLocales,
+      activeLocales: props.activeLocales,
     };
 
     const { hasCardEditActions, hasCardMoveActions, hasCardRemoveActions } = props;