+++ packages/date/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.6](https://github.com/contentful/field-editors/compare/@contentful/field-editor-date@1.7.5...@contentful/field-editor-date@1.7.6) (2025-04-30)
+
+**Note:** Version bump only for package @contentful/field-editor-date
+
 ## [1.7.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-date@1.7.4...@contentful/field-editor-date@1.7.5) (2025-04-03)
 
 **Note:** Version bump only for package @contentful/field-editor-date