+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.19](https://github.com/contentful/field-editors/compare/@contentful/field-editor-radio@1.5.18...@contentful/field-editor-radio@1.5.19) (2025-04-30)
+
+**Note:** Version bump only for package @contentful/field-editor-radio
+
+## [1.5.18](https://github.com/contentful/field-editors/compare/@contentful/field-editor-radio@1.5.17...@contentful/field-editor-radio@1.5.18) (2025-04-24)
+
+**Note:** Version bump only for package @contentful/field-editor-radio
+
 ## [1.5.17](https://github.com/contentful/field-editors/compare/@contentful/field-editor-radio@1.5.16...@contentful/field-editor-radio@1.5.17) (2025-04-03)
 
 **Note:** Version bump only for package @contentful/field-editor-radio