+++ packages/single-line/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-single-line@1.6.0...@contentful/field-editor-single-line@1.6.1) (2025-04-30)
+
+**Note:** Version bump only for package @contentful/field-editor-single-line
+
+# [1.6.0](https://github.com/contentful/field-editors/compare/@contentful/field-editor-single-line@1.5.5...@contentful/field-editor-single-line@1.6.0) (2025-04-24)
+
+### Features
+
+- Improve Accessibility of field editors by enabling associating them with labels [UFO-1737] ([#1856](https://github.com/contentful/field-editors/issues/1856)) ([914f643](https://github.com/contentful/field-editors/commit/914f643bb93cbe523a641ad5430e7399fe71987b))
+
 ## [1.5.5](https://github.com/contentful/field-editors/compare/@contentful/field-editor-single-line@1.5.4...@contentful/field-editor-single-line@1.5.5) (2025-04-03)
 
 **Note:** Version bump only for package @contentful/field-editor-single-line